乐观锁工具类修复

+++ icloud-starters/icloud-starter-drools/src/main/java/cn/isqing/icloud/starter/drools/dao/entity/RunLog.java
@@ -24,7 +24,7 @@ public class RunLog {
 
     private Integer lockStatus;
 
-    private Long lockVersion;
+    private Integer lockVersion;
 
     private LocalDateTime lockTime;
